:truck: Reorganise the packages
